Fix previewText message display when unchecked
